# service1-ProductService/api-specs/api.yaml
openapi: 3.0.1
info:
  title: Product Service API
  description: API documentation for Product Microservice
  version: "1.0"
servers:
- url: http://localhost:8081
  description: Generated server url
tags:
- name: Product API
  description: Operations related to products
paths:
  /products/{id}:
    get:
      tags:
      - Product API
      summary: Get product by ID
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
    put:
      tags:
      - Product API
      summary: Update product
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      tags:
      - Product API
      summary: Delete product
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /products/reduceStock/{id}/{qty}:
    put:
      tags:
      - Product API
      summary: Reduce product stock
      description: Used by Order-Service via FEIGN client
      operationId: reduceStock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: qty
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /products:
    get:
      tags:
      - Product API
      summary: Get all products
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags:
      - Product API
      summary: Add new product
      operationId: add
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          format: int32